# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "RDFConverter",
    "version": "1.0.0",
    "title": "RDFConverter",
    "contact": {
      "name": "Thomas Hanke",
      "email": "thomas.hanke@imws.fraunhofer.de"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/api/yarrrmltorml": {
      "post": {
        "tags": [
          "YARRRML to RML"
        ],
        "summary": "Convert YARRRML to RML",
        "description": "Given an YARRRML file url, it generates RDF containing the RML rules.",
        "requestBody": {
          "required": "true",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Url"
              }
            }
          }
        },
        "produces": [
          "text/plain"
        ],
        "responses": {
          "200": {
            "description": "RDF data containing the RML rules."
          },
          "404": {
            "description": "File not found."
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/joindata": {
      "post": {
        "tags": [
          "Join Data"
        ],
        "summary": "Apply the RML mapping to the data.",
        "description": "Apply the RML mapping specified in the rml_url to the data_url.",
        "parameters": [
          {
            "in": "body",
            "name": "rml_url",
            "type": "string",
            "format": "uri",
            "required": "true",
            "description": "URL to a file containing an RML Mapping"
          },
          {
            "in": "body",
            "name": "data_url",
            "type": "string",
            "format": "uri",
            "required": "true",
            "description": "URL to RDF data to which the mapping should be applied."
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Mapped RDF data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinData"
                }
              }
            }
          },
          "404": {
            "description": "File not found."
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/rdfvalidator": {
      "post": {
        "tags": [
          "Validate RDF against SHACL"
        ],
        "summary": "Validate some RDF data against SHACL shapes.",
        "description": "Validate some RDF data against SHACL shapes.",
        "parameters": [
          {
            "in": "body",
            "name": "shapes_url",
            "type": "string",
            "format": "uri",
            "required": "true",
            "description": "URL to a file containing SHACL shapes."
          },
          {
            "in": "body",
            "name": "rdf_url",
            "type": "string",
            "format": "uri",
            "required": "true",
            "description": "URL to RDF data which should be validated against the SHACL shapes."
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Json containing the Validation report and a boolean variable indicating whether the RDF conforms to she SHCAL shapes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/shacl_report"
                }
              }
            }
          },
          "404": {
            "description": "File not found."
          },
          "400": {
            "description": "The validation has produced an error."
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Url":{
        "description": "Conversion object handling url to files",
        "required": "true",
        "type": "object",
        "properties": {
          "url":{
            "type": "string",
            "format": "uri",
            "description": "Url to yarrrml file",
            "example": "https://github.com/Mat-O-Lab/MapToMethod/raw/main/examples/example-map.yaml"
          }
        }
      },
      "shacl_report": {
        "description": "Report of SHACL validation.",
        "required": "true",
        "type": "object",
        "properties": {
          "valid": {
            "type": "boolean",
            "description": "True if RDF is valid according to SHACL shape."
          },
          "graph": {
            "type": "string",
            "description": "RDF triples describing the validation report of the SHACL validation."
          }
        }
      },
      "JoinData": {
        "description": "Mapped RDF Data and count of applied or skipped rules",
        "type": "object",
        "properties": {
          "graph": {
            "type": "string",
            "description": "Mapped RDF Graph in turtle format."
          },
          "num_mappings_applied": {
            "type": "integer",
            "description": "Count of mappings from data to method graph that have been applied."
          },
          "num_mappings_skipped": {
            "type": "integer",
            "description": "Count of mappings from data to method graph that have been skipped because they could not be applied."
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
    "ApiResponse": {
      "type": "object",
      "properties": {
        "filename": {
          "type": "string"
        },
        "filedata": {
          "type": "string"
        }
      }
    }
  }
}